Update to ares v118r15 release.
[This release was posted to the ares Discord after r16. That version introduced
128KB VRAM mode for the Super Famicom core, and somebody requested the previous
version so they could see the changes involved in implementing that feature.
-Ed.]

+++ ares/component/processor/sh2/instructions.cpp
@@ -179,12 +179,12 @@ auto SH2::DIV1(u32 m, u32 n) -> void {
 
 //DMULS.L Rm,Rn
 auto SH2::DMULS(u32 m, u32 n) -> void {
-  MAC = (s32)R[n] * (s32)R[m];
+  MAC = (s64)(s32)R[n] * (s32)R[m];
 }
 
 //DMULU.L Rm,Rn
 auto SH2::DMULU(u32 m, u32 n) -> void {
-  MAC = R[n] * R[m];
+  MAC = (u64)R[n] * R[m];
 }
 
 //DT Rn
